requirements.txt
requests
pytest
pyOpenSSL